rework wrapper to contain the model points and colors, loop in JS land
this seems to speed it up a bit. Its faster to loop in JS than poke
variables from Java.
This makes it a bit easier to partition off a whole wrapper context to
a subset of some LX points. As a crazy demo of this,
PixelblazePatternParallel runs a ton of these across a number of
concurrent threads.

There's definitely some chunkiness happening on first startup, but once
the JIT/memory warms up it seems smooth.