Add fire pattern (#31)
Forgot to port this over when we switched repos. This is now connected to our
palette! It looks a little weird just lerp-ing the color from black -> [palette
color] -> white (when the palette color is purple, there's a lot of green in the
result for example), so I may want to add some additional parameters.