Implement palettes discussed (#108)
Note this reorders the indices of palettes to better align with LX's
comprehension of swatch order. Project files that use a LinkedColorParameter
may need to reset their palette index. For example, a foreground used to be
linked to UI index (1-based) 1, and would now need to be linked to 3 (the
new primary). Patterns that made use of color constants or helpers should
be unaffected.