Edge and panel IP assignments - alt versions
Commited by Jeff, but these unknown author edits look to me
to be by someone who needed to perhaps silence LX complaining
of controller endpoints not found, or perhaps as a way to
limit which controllers LX was attempting to communicate with.

If the person who made these changes sees this, please consider
ammending this commit message or reverting this commit.