Beat-linked speed & Angle control are in.
Angle control, which allows you to set a static rotation angle for a pattern, is in.  It is linked to spin, the pattern always returns to the set static angle when spin is reset or turned off.

The speed control is now tied to the beat - engine bpm changes will be reflected in pattern speed, and the control acts as a bpm multipler,  with a default range of double whole notes to 16th notes at the current tempo.   There is a retrigger() API to allow precise sync to beat and measure start.