Brightness rework - Built into color control
The Brightness control is now automatically used to adjust the output of the color control, so pattern developers do not have to worry about multiplying color by brightness in patterns.  If you get color from the common control, it will just work.  This applies to both Java and GLSL.

(Also sneaking in the ArcEdges pattern - controls are definitely not tuned yet, but I'd like to get some feedback on how it runs on a Mac.  If it doesn't blow up because I've missed removing some nvidia only feature, I'd expect M1/M2 machines to be fine.  Might slow down the Intel models a bit.)