Tools for Updating FragmentShaderEffect Patterns (#199)
* Tools for Updating FragmentShaderEffect Patterns

A few changes needed to simplify converting these patterns to use the common controls, plus a sample conversion of the NeonSnakeShader pattern.

This is a straightforward, very vanilla conversion, showing how to set up and use the common controls in this pattern framework.  We can expand this pattern later if we want to.  Controls are:

- Color: Yes
- Speed: Yes
- X/Y Offset: Y only, since the pattern moves constantly in X
- Spin/Angle:  No.
- Brightness: Yes
- Size:  Snake size
- Quantity:  No
- Wow1:  Base Glow
- Wow2:  Beat reactive snake size modulation
- Wow Trigger: No

* Performance enhancements, cleanup, rotation support

* Color control fixes for Pixelblaze javascript patterns

Needed to keep current AutoVJ from dropping to 1fpswhen running a native PB pattern -- it uses a couple and they were causing log flooding with "don't use getPrimary/SecondaryGradientColor()" messages.

Updated the js framework to use getGradientColor()

* Formatting: reduce default spacing on continued lines

8 seems kind of excessive.

* Formatting cleanup

---------

Co-authored-by: Justin Belcher <jkbelcher@gmail.com>