Take the gray out of gradients (#201)
Thanks to upstream assistance from mcslee, future LX releases will have improved color blending especially when blending to/from black.

This PR brings in the new blend functions so we can use them immediately. This is a temporary patch until we upgrade to a version of LX that includes the new features.