New noise patterns, Round 1
These two patterns show the general direction I'm taking with noise.  Rather than writing one gigantic
noise pattern to rule them all, I'm doing an assortment of small, high performance patterns, with features that work with to the specific type and appearance of the noise.

When testing, please be sure to turn off any LX modulators that are active on the channel. No problem using them in production if necessary, but they do tend to ~~stomp~~ obscure pattern detail.

### Turbulence Lines
Generates a value noise based turbulence field, then uses the field
to draw sinuous, organic-looking lines.  Wow1 controls the noise field/line mix,
Wow2 controls the foreground color/gradient mix.

Wow trigger fires a double pulse on the beat that lights the lines (not the background fog),
so you can turn the lines way down in the mix, and they will be revealed as the pulse passes.

- Speed: Yes
- X/Y Offset: Yes
- Spin/Angle: Yes
- Brightness: Yes
- Size: Overall scale
- Quantity: Line density
- Wow1: Noise field/wavy line mix
- Wow2: Foreground color/gradient mix
- WowTrigger: Radial double pulse on the beat

### Simplex Posterized
Generates a cloudy simplex noise field which you can color with the current gradient,
and "Posterize" - quantize to a small number of brightness levels - for a cartoon-like look.

Since static x and y offset are not useful for this sort of pattern, they are instead used to
control "drift" - x and y movement speed.

Wow1 controls the posterization level, Wow2 controls the foreground color/gradient mix.

- Speed: Yes
- X/Y Offset: Yes, controls "drift" speed and directrion
- Spin/Angle: Yes
- Brightness: Yes
- Size: Overall scale
- Quantity: Noise field density
- Wow1: Posterization level
- Wow2: Foreground color/gradient mix
- WowTrigger: No