Rotation & scaling in TEGradientPattern (#233)
Fixed -- LXParameterizedMatrix seems to want its roll, pitch and yaw to also be CompoundParameters, which we could do, but seems a little awkward for this job.

Switched to LXMatrix, and added a call to identity() to reset the transform at the start of every frame b/c all our controls give absolute angles and offsets rather than incremental.

Let me know if there's anything else this needs.  Given how sparse our palettes are, we might still need to play with the parameterization a bit for best visuals.