Fix "Blast Radius" bug in MatrixScroller
This broke when we switched to Views because it was originally written to expect a much larger coordinate range.  To see the bug, turn up Wow1 on MatrixScroller in the current main.

Reworked the random dispersion math a little for the new coordinate range, and to give a smoother pattern.