Fix for NDI send failure after pattern load (#466)
One line fix:  Need to be sure that the channel name is removed when the effect is disposed.   The GC will take care of this eventually, but since it involves native memory, "eventually" might be a very long time.

Since we're using one channel name for all TE NDI output, explicitly closing the channel on dispose frees that name for immediate re-use by the next loaded pattern.