Move beat reactive size calc
Try this and see if you like it!

I moved the beat reactive size calculation (nice!) out of the blob accumulator loop, since it's per-frame invariant anyway, and set it up so that the pattern won't drop to black if there's no audio signal.

Also removed the references to the new EnterTheMatrix pattern from TEApp.java, since the actual code isn't in this PR.