Restore Panel Backings and Labels (#514)
* Fix model axes for HandTracker pattern

Should work normally now on both static and dynamic models.

Also fixed a stray vocal/vocals stem-related variable name that was hiding in the shader framework.  (And changed it in the docs, too.)

* Checkpoint - panel labels ok

* Backings are back

working on both static and dynamic model

* Fixes from code review

- General cleanup
- Make 3D Ui components thread safe.

* Correct really messed-up commit

I think I accidentally discarded some changes that should've been checked in...

* Remove unnecessary diagnostics

* more cleanup again

...stoopid System.out.print()

* Fix long-standing resource leak

Thank you @jkbelcher for putting me on the path to finding this!