WIP: Fixture Cleanup (#511)
* fix reverse on edge outputs

* Fix model axes for HandTracker pattern

Should work normally now on both static and dynamic models.

Also fixed a stray vocal/vocals stem-related variable name that was hiding in the shader framework.  (And changed it in the docs, too.)

* fix json5 syntax in panel files

* fix edge artnet output

* Add legacy Edge and Panel tags to model conversion script (#513)

* Fix model axes for HandTracker pattern (#512)

Should work normally now on both static and dynamic models.

Also fixed a stray vocal/vocals stem-related variable name that was hiding in the shader framework.  (And changed it in the docs, too.)

* Add legacy Edge and Panel tags to model conversion script

---------

Co-authored-by: ZRanger1 <35553094+zranger1@users.noreply.github.com>

* Remove .lxpreferences from branch

---------

Co-authored-by: ZRanger1 <35553094+zranger1@users.noreply.github.com>
Co-authored-by: Justin Belcher <jkbelcher@gmail.com>