fix(SimplifyEffect): add check before this.getModelView() (#688)
* fix(SimplifyEffect): add check before this.getModelView()

When `SimplifyEffect` was calling `getModelView()` (right after being added to a channel), we were seeing an exception get thrown. Specifically here

```java
  // In LXDeviceComponent:

  public final LXModel getModelView() {
    LXViewDefinition view = this.view.getObject();
    if (view != null) {
      return view.getModelView();
    }
    return switch (getParent()) {  // <--- Exception thrown here
      case LXMasterBus master -> lx.model;
      case LXAbstractChannel bus -> bus.getModelView();
      case LXPattern pattern -> pattern.getModelView();
      default -> getModel();
    };
  }
```

It seems like maybe upstream LX should do a null check on that `getParent()` call, just to be safe?

I'm not totally sure where in the lifecycle that parent gets set on an LXEffect - maybe there's a better way for this effect to decide when to first refresh its models.

It seems to me like there are other places that `refreshModels` gets called. If `depth` defaults to `0` (no visual impact), and turning up the depth invokes a param listener that will call `refreshModels`, I think this is safe. Works for me when I tested locally.

* getModelViewFixed()

* PR feedback